Removed dublicate tests
